Move libs to separate directory
