+++ lib/metalsmith-plugins.js
