+++ lib/travis-utils.js
